change to doubles

+++ src/main/java/frc/robot/utilities/HowdyMath.java
@@ -5,9 +5,9 @@ public class HowdyMath {
   // with respect to m using extended
   // Euclid Algorithm. Refer below post for details:
   //
-  public static int inverse_modulus(int a, int m) {
-    int m0 = m, t, q;
-    int x0 = 0, x1 = 1;
+  public static double inverse_modulus(double a, double m) {
+    double m0 = m, t, q;
+    double x0 = 0, x1 = 1;
     if (m == 1) return 0;
 
     // Apply extended Euclid Algorithm
